# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StitchFlow API",
    "version": "1.1.0",
    "description": "The contract between the StitchFlow apps, the web admin panel and the API. Every operation names the permission action it needs (x-action); a destructive action is super admin only. Dates without a time are \"YYYY-MM-DD\"; instants are ISO-8601 UTC; money is rounded to the paisa. See HANDOVER.md."
  },
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Replace per environment."
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": [
          "superAdmin",
          "admin",
          "viewAdmin",
          "supervisor",
          "worker"
        ],
        "description": "The five fixed roles. `worker` is shown as Operator."
      },
...
      "ShiftScheme": {
        "type": "string",
        "enum": [
          "dayNight",
          "threeShift",
          "manual"
        ]
      },
...
      "AccessLevel": {
        "type": "string",
        "enum": [
          "none",
          "view",
          "edit"
        ],
        "description": "none < view < edit."
      },
      "AccountStatus": {
        "type": "string",
        "enum": [
          "active",
          "disabled"
        ]
      },
      "PropertyStatus": {
        "type": "string",
        "enum": [
          "active",
          "suspended",
          "closed"
        ]
      },
...
      "ModuleKey": {
        "type": "string",
        "enum": [
          "production",
          "attendance",
          "advances",
          "payroll",
          "staff",
          "machines",
          "shifts",
          "bonus_rules",
          "recycle_bin",
          "users",
          "settings"
        ],
        "description": "From the SQL catalog."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "validation_failed",
          "unauthenticated",
          "token_expired",
          "forbidden",
          "super_admin_only",
          "owner_only",
          "password_change_required",
          "property_suspended",
          "not_found",
          "machine_number_taken",
          "machine_limit_reached",
          "staff_code_taken",
          "period_already_paid",
          "restore_blocked",
          "shift_cycle_clash",
          "slab_ladder_invalid",
          "grant_exceeds_granter",
          "file_too_large",
          "rate_limited"
        ]
      },
      "Property": {
        "type": "object",
        "description": "One customer's unit — the tenant.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "code": {
            "type": "string",
            "description": "Short handle, e.g. \"janki-creation\"."
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string",
            "description": "IANA zone; \"today\" is this zone's date."
          },
          "status": {
            "$ref": "#/components/schemas/PropertyStatus"
          },
          "machineLimit": {
            "type": "integer",
            "description": "Most machines the unit may have. Set from the web admin panel only."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "timezone",
          "status",
          "machineLimit",
          "createdAt"
        ]
      },
      "PropertySettings": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string"
          },
          "currencySymbol": {
            "type": "string"
          },
          "shiftScheme": {
            "$ref": "#/components/schemas/ShiftScheme"
          },
          "dayStartMinute": {
            "type": "integer",
            "minimum": 0,
            "maximum": 1439
          },
          "noLeaveBonusAmount": {
            "type": "number",
            "multipleOf": 0.01,
            "description": "Rupees, two decimals."
          },
          "stitchBasedEnabled": {
            "type": "boolean"
          },
          "stitchBasedRatePerThousand": {
            "type": "number",
            "description": "Up to four decimals."
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "businessName",
          "currencySymbol",
          "shiftScheme",
          "dayStartMinute",
          "noLeaveBonusAmount",
          "stitchBasedEnabled",
          "stitchBasedRatePerThousand",
          "timezone"
        ]
      },
      "User": {
        "type": "object",
        "description": "A person who can sign in. Never carries a password hash.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "displayName": {
            "type": "string"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "description": "E.164, e.g. \"+919812345678\"."
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "mustChangePassword": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "displayName",
          "username",
          "phone",
          "email",
          "status",
          "mustChangePassword",
          "lastLoginAt"
        ]
      },
...
      "ModuleAccess": {
        "type": "object",
        "properties": {
          "module": {
            "$ref": "#/components/schemas/ModuleKey"
          },
          "level": {
            "$ref": "#/components/schemas/AccessLevel"
          }
        },
        "required": [
          "module",
          "level"
        ]
      },
      "Member": {
        "type": "object",
        "description": "A user's place in a property.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "Server-issued id."
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "isOwner": {
            "type": "boolean",
            "description": "The property's main super admin."
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "staffId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The roster row this login is, if any."
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleAccess"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "displayName",
          "role",
          "isOwner",
          "status",
          "staffId",
          "modules",
          "createdAt"
        ]
      },
      "Me": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "settings": {
            "$ref": "#/components/schemas/PropertySettings"
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleAccess"
            },
            "description": "Gate every screen on this list."
          }
        },
        "required": [
          "user",
          "member",
          "property",
          "settings",
          "modules"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "description": "English, for logs. Clients word `code` in the user's language."
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "PropertyPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PropertyStatus"
          },
          "machineLimit": {
            "type": "integer",
            "description": "Lowering it below the machines the unit has removes nothing; it only stops adding."
          }
        }
      },
...
      "PropertyDetail": {
        "type": "object",
        "properties": {
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "settings": {
            "$ref": "#/components/schemas/PropertySettings"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          },
          "machineCount": {
            "type": "integer"
          },
          "staffCount": {
            "type": "integer"
          }
        },
        "required": [
          "property",
          "settings",
          "members",
          "machineCount",
          "staffCount"
        ]
      },
...
    }
  },
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "summary": "Who is signed in, where, and with which modules.",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/platform/properties/{id}": {
      "get": {
        "operationId": "getPlatformPropertiesById",
        "summary": "One property.",
        "tags": [
          "platform/properties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetail"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-platform-role": "support"
      },
      "patch": {
        "operationId": "patchPlatformPropertiesById",
        "summary": "Rename, change zone, set the machine limit, suspend or reactivate.",
        "tags": [
          "platform/properties"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PropertyPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            }
          },
          "default": {
            "description": "An error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-platform-role": "admin"
      }
    },
...
  }
}